Rename existing theme files
